LangChain/02 LangChain-Practical/01 LangChain Models/3.EmbeddedModels/2_embedding_openai_docs.py
from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv

load_dotenv()

embeddings = OpenAIEmbeddings(model = "text-embedding-3-large", dimensions=32)

documents = [
    "Delhi is the capital of India",
    "Bangalore is the capital of Karnataka",
    "Paris is the capital of France"
]

result = embeddings.embed_documents(documents)

print(str(result))